fix(index.ts): determine root folder at runtime to properly sanitize absolute paths
Currently lint files absolute paths are sanitized by removing the current folder from there. It is
however possible that the current folder does not match the folder in the lint file. For instance,
if the lint file was generated ni a different machine than where danger is running. This is common
when using dockerized instances and performing lint in one stage and danger in a follow-up one.