fix: support lint files calculated in a different machine/process (#31)
When building in CI it is common that a task is generating the lint file, while a different task
runs dagger. These tasks don't need to run in the same machine. The current code assumes so when
sanitizing absolute paths by replacing the root (pwd) with "". This does not work if the root is
different.

Closes #29

Co-authored-by: Miguel Cohnen <miguel@ixigo.com>